BaseS3BucketEngine: new class
Also: refactor in preparation for CephS3BucketEngine
