+++ rfi_file_monitor/operations/dummy_operation.py
@@ -11,12 +11,12 @@ from random import random
 from ..operation import Operation
 from ..utils.exceptions import SkippedOperation
 from ..utils.decorators import with_pango_docs, supported_filetypes, add_directory_support
-from ..file import File, RegularFile, AWSS3Object, URL, Directory
+from ..file import File, RegularFile, S3Object, URL, Directory
 
 logger = logging.getLogger(__name__)
 
 @with_pango_docs(filename='dummy_operation.pango')
-@supported_filetypes(filetypes=[RegularFile, AWSS3Object, URL, Directory])
+@supported_filetypes(filetypes=[RegularFile, S3Object, URL, Directory])
 class DummyOperation(Operation):
 
     NAME = "Dummy Operation"
